feat(styling): refactor styling setup
